% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_load}
\alias{catalog_load}
\title{Load dataset from configuration}
\usage{
catalog_load(dataset_name)
}
\arguments{
\item{dataset_name}{The name of the dataset to load as specified in the catalog}
}
\value{
The loaded dataset as a data frame
}
\description{
This function reads a catalog configuration from a YAML file and loads the specified
dataset based on its configuration. Currently supports SQL queries from Snowflake,
as well as CSV and Excel files.
}
\examples{
\dontrun{
df <- catalog_load("example_data")
}
}
