% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forge.R
\name{forge}
\alias{forge}
\title{Create a new project directory structure}
\usage{
forge()
}
\value{
A message indicating that the project structure was created successfully.
If a directory with the specified project name already exists, the function
stops and returns an error message.
}
\description{
This function asks the user for a project name and creates a new directory
structure for the project, inspired by Kedro. Key directories include 'conf', 'data',
'src', 'logs', 'notebooks', and 'docs'. It pre-populates 'conf/base' with
'catalog.yaml' and 'parameters.yaml', 'conf/credentials' with 'credentials.yaml',
and 'src' with '\emph{main}.R', 'nodes.R', and 'pipelines.R' templates to demonstrate
a runnable example pipeline.
}
\examples{
\dontrun{
forge()
}
}
