% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parameters.R
\name{load_parameters}
\alias{load_parameters}
\title{Load parameters from a YAML file}
\usage{
load_parameters(path = "conf/base/parameters.yaml")
}
\arguments{
\item{path}{A character string specifying the path to the parameters YAML file.
Defaults to "conf/base/parameters.yaml".}
}
\value{
A list containing the parameters loaded from the YAML file. If the
file is not found or cannot be parsed, the function will stop with an error.
}
\description{
This function reads parameters from a specified YAML file. These parameters
can then be used to configure nodes and pipelines.
}
\examples{
\dontrun{
# params <- load_parameters()
# custom_params <- load_parameters("conf/custom/my_params.yaml")

# To use in a node, you might pass parameter names:
# create_node(
#   func = my_func,
#   inputs = "input_data",
#   outputs = "output_data",
#   parameters = list(value = "params_key_from_yaml")
# )
# And then the pipeline runner would resolve "params_key_from_yaml"
# against the loaded parameters.
}
}
