% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orchestrate.R
\name{orchestrate}
\alias{orchestrate}
\title{Run the '\emph{main}.R' script}
\usage{
orchestrate()
}
\value{
None
}
\description{
This function runs the '\emph{main}.R' file located in the 'src' directory. This is part of the project's
orchestration process, enabling the execution of a predefined workflow.
Note: This function assumes that the working directory is set correctly.
}
\examples{
\dontrun{
orchestrate()
}
}
