% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/catalog.R
\name{catalog_load}
\alias{catalog_load}
\title{Load dataset from configuration}
\usage{
catalog_load(
  dataset_name,
  catalog_path = "conf/base/catalog.yaml",
  credentials_path = "conf/credentials/credentials.yaml"
)
}
\arguments{
\item{dataset_name}{The name of the dataset to load as specified in the catalog}

\item{catalog_path}{The path to the catalog YAML file. Defaults to "conf/base/catalog.yaml".}

\item{credentials_path}{The path to the credentials YAML file. Defaults to "conf/credentials/credentials.yaml".}
}
\value{
The loaded dataset as a data frame
}
\description{
This function reads a catalog configuration from a YAML file and loads the specified
dataset based on its configuration. Currently supports SQL queries from Snowflake,
as well as CSV and Excel files.
}
\examples{
\dontrun{
df <- catalog_load("example_data")
# df_custom_paths <- catalog_load("example_data",
#                                 catalog_path = "conf/custom/catalog.yaml",
#                                 credentials_path = "conf/custom/credentials.yaml")
}
}
